+++ pkg/kubelet/kubelet_test.go
@@ -113,7 +113,7 @@ func newTestKubelet(t *testing.T) *TestKubelet {
 	podManager, fakeMirrorClient := newFakePodManager()
 	kubelet.podManager = podManager
 	kubelet.containerRefManager = kubecontainer.NewRefManager()
-	kubelet.containerRunner = &dockertools.DockerContainerRunner{fakeDocker, fakeRecorder}
+	kubelet.containerManager = dockertools.NewDockerManager(fakeDocker, fakeRecorder)
 	return &TestKubelet{kubelet, fakeDocker, mockCadvisor, fakeKubeClient, waitGroup, fakeMirrorClient}
 }
